refactor: move Migrate method from datastore to prepare microservice to migrations
Changelog: None
